refactor(abstraction): modify structure and add person type
still missing test modifications

fix #1787
