+++ src/model/Patient.ts
@@ -1,7 +1,6 @@
-import RelatedPerson from './RelatedPerson'
 import Person from './Person'
 
 export default interface Patient extends Person {
   friendlyId: string
-  relatedPersons?: RelatedPerson[]
+  firstVisit?: boolean
 }
