+++ src/persons/patients/new/NewPatientForm.tsx
@@ -2,11 +2,11 @@ import React, { useState } from 'react'
 import { useTranslation } from 'react-i18next'
 import { Alert, Button, Checkbox } from '@hospitalrun/components'
 import { startOfDay, subYears } from 'date-fns'
-import SelectWithLabelFormGroup from '../../components/input/SelectWithLableFormGroup'
-import TextFieldWithLabelFormGroup from '../../components/input/TextFieldWithLabelFormGroup'
-import TextInputWithLabelFormGroup from '../../components/input/TextInputWithLabelFormGroup'
-import DatePickerWithLabelFormGroup from '../../components/input/DatePickerWithLabelFormGroup'
-import Patient from '../../model/Patient'
+import SelectWithLabelFormGroup from '../../../components/input/SelectWithLableFormGroup'
+import TextFieldWithLabelFormGroup from '../../../components/input/TextFieldWithLabelFormGroup'
+import TextInputWithLabelFormGroup from '../../../components/input/TextInputWithLabelFormGroup'
+import DatePickerWithLabelFormGroup from '../../../components/input/DatePickerWithLabelFormGroup'
+import Patient from '../../../model/Patient'
 import { getPatientName } from '../util/patient-name-util'
 
 interface Props {
